FutStats v0.0.1 FutStats.Utils

Utilized by the seed.exs for seeding the database. Reads the Data.csv file as a stream, maps each record to the corresponding Ecto Schema and inserts it into the database.

Functions

insert_record(record)

Converts the map obtained from a csv line to an Ecto.Schema

Parameters

record: map representing a csv line

insert_records(csv_stream)

Accepts a File.Stream object from the seed.exs script and calls insert_record/1 for each entry.

Parameters

csv_stream: A File.Stream