FutStats v0.0.1 FutStatsWeb.MatchController

Controller for /matches_for_league_season route. Contains only a single method matches_for_league_season/2 which send the matches for a particular league season pair as response. Calls FutStats.Queries.Matches.list_matches_for_league_season/2 Accepts 3 parameters in the query string, league, season and format

Functions

matches_for_league_season(conn, map)

swagger_definitions()

swagger_path_matches_for_league_season(route)